% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z scores.R
\name{my_z}
\alias{my_z}
\title{Make a vector of Z scores}
\usage{
my_z(x)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
A vector of z values
}
\description{
Make a vector of Z scores
}
\examples{
my_z(1:4)

}
